20220423 1.調整後端傳至前端的方式 不使用ViewData 2.AccountSelectDto Level型態byte改成int

+++ OnlineShopBack/DTOs/AccountSelectDto.cs
@@ -8,7 +8,7 @@ namespace OnlineShopBack.Controllers
         public int Id { get; set; }
         public string Account { get; set; }
         public string Pwd { get; set; }
-        public byte Level { get; set; }
+        public int Level { get; set; }
 
         public DateTime CreateDate { get; set; }
     }
